move application calls to ssc.ts
